add more benchmark results; refactor the code a bit and remove more template parameters
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/AndreasKirsch@713 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/cuda/config.h
@@ -94,7 +94,7 @@
 #	define MAX_REGISTER_COUNT 63
 #endif
 
-#define BLOCK_SIZE (WARP_SIZE*NUM_WARPS)
+#define BLOCK_SIZE uint(WARP_SIZE*NUM_WARPS)
 
 #ifdef CUDA_DOUBLE_MODE
 	typedef double floatType;
